finished version1 of homework6
